dags/transform/requirements.txt
PyExifTool==0.5.5
pandas
python-dotenv==0.21.1